requirements_mac.txt
aiohttp==3.9.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
effdet==0.4.1
emoji==2.11.0
et-xmlfile==1.1.0
etelemetry==0.3.1
Events==0.5
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.1
fastjsonschema==2.19.1
filelock==3.13.4
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
iopath==0.1.10
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.25
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.2
langchain-openai==0.1.8
langchain-text-splitters==0.2.0
langchainhub==0.1.15
langdetect==1.0.9
langserve==0.0.51
langsmith==0.1.45
layoutparser==0.3.4
looseversion==1.3.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.15.1
openai==1.30.5
opencv-python==4.9.0.80
openpyxl==3.1.2
opensearch-py==2.6.0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
pikepdf==8.15.0
pillow==10.3.0
pillow_heif==0.16.0
platformdirs==4.2.1
portalocker==2.8.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
prov==2.0.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.0
pydot==2.0.0
Pygments==2.17.2
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.29.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.9
pytube==15.0.0
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
rapidfuzz==3.8.1
rdflib==7.0.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
Send2Trash==1.8.3
sentence-transformers==2.6.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sse-starlette==2.1.0
stack-data==0.6.3
starlette==0.37.2
streamlit==1.33.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.4.0
tiktoken==0.7.0
timm==0.9.16
tinycss2==1.3.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
traits==6.3.2
transformers==4.39.3
types-python-dateutil==2.9.0.20240316
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
unstructured==0.13.2
unstructured-client==0.18.0
unstructured-inference==0.7.25
unstructured.pytesseract==0.3.12
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.29.0
watchdog==4.0.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.10
wrapt==1.16.0
yarl==1.9.4
youtube-transcript-api==0.6.2
